Refactor package structure for authentication and user account classes
