--- src/main/java/com/guia747/accounts/Entity.java
@@ -1,12 +1,0 @@
-package com.guia747.accounts;
-
-import java.util.UUID;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class Entity {
-
-    private UUID id;
-}
