--- src/main/java/com/guia747/accounts/UserAccountService.java
@@ -1,8 +1,0 @@
-package com.guia747.accounts;
-
-import com.guia747.authentication.OAuth2UserProfile;
-
-public interface UserAccountService {
-
-    UserAccount findOrCreateFromOAuth2(OAuth2UserProfile oauth2User);
-}
