+++ src/main/java/com/guia747/authentication/dto/OAuth2AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.guia747.authentication;
+package com.guia747.authentication.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import io.swagger.v3.oas.annotations.media.Schema;
