+++ src/main/java/com/guia747/authentication/usecase/OAuth2AuthenticationUseCase.java
@@ -1,7 +1,7 @@
 package com.guia747.authentication.usecase;
 
-import com.guia747.accounts.UserAccount;
-import com.guia747.authentication.OAuth2AuthenticationRequest;
+import com.guia747.accounts.domain.UserAccount;
+import com.guia747.authentication.dto.OAuth2AuthenticationRequest;
 
 public interface OAuth2AuthenticationUseCase {
