+++ src/main/java/com/guia747/infrastructure/persistence/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package com.guia747.infrastructure.persistence.mapper;
 
 import org.mapstruct.Mapper;
-import com.guia747.accounts.UserAccount;
+import com.guia747.accounts.domain.UserAccount;
 import com.guia747.infrastructure.config.GlobalMapperConfig;
 import com.guia747.infrastructure.persistence.jpa.entity.UserJpaEntity;
